Add "not coming" emoji to "React with..." message
Modified the "React with..." line on the event reminder message to specify an emoji to react with if you're not coming to the meeting.

The current solution to specifying if you're not able to come to an event is by posting a message or replying to the thread in the reminder message. This should make it easier to know how many people will not be coming to the meeting at a glance.