Create independent function for emoji generation
Modified the generateMessage function to accept emojis as a parameter and removed the code that generates the emoji pair for the coming/not coming reactions in the message. Instead, a function has been created that returns a pair of non-duplicate emojis for use by generateMessage. The generateMessage call in the checkForEvents function has been modified to have the emojis parameter, but emojis will only be passed if the message being generated required them.

Since the emojis used in generated messages will soon need to be known by functions other than generateMessage (e.g. for reaction seeding), a function to generate emojis has been created and parameters in generateMessage modified to accept emojis through this function