Modify getEmojiPair to return set emojis on fail
The getEmojiPair function has been modified so that instead of rejecting the promise after five attempts at getting a unique emoji pair, it will return a pair of hard-coded emojis: [":white_check_mark:", ":x:"] currently.

While the chance of the function failing when the Slack API is running is astronomically low, this implementation will ensure that the function will also function if the API fails to retrieve emojis.