Implement "React if not coming" emoji to reminder message  (#15)
* Fix test cases

* Format spaces to be all the same type (thanks jacob)

* Update firebase

* Handle missing agenda and notes lines

* Change from substring to string search

* Add test for multiple channels

* Fix logic for extra lines

* Address PR comments

* Improve logging

* Fix typo

* Update datetime to better format

* Change to moment-timezone since intl doesn't work with timezones

* Stop links from unfurling unnecessarily

* Change to new google account structure

* update npm

* Handle undefined

* Fix mocking wrong value

* Handle empty descriptions better

* Add more channels to the default option

* Add Jetbrains IDE files to .gitignore

* Add "not coming" emoji to "React with..." message

Modified the "React with..." line on the event reminder message to specify an emoji to react with if you're not coming to the meeting.

The current solution to specifying if you're not able to come to an event is by posting a message or replying to the thread in the reminder message. This should make it easier to know how many people will not be coming to the meeting at a glance.

* Modify emoji getter to better handle duplicates

Modified the emoji getter for generateMessage to handle receiving duplicate emojis more effectively by retrying the emoji request five times and then rejecting the promise.

Previously, the emoji getter was on a while loop so the function could have potentially run forever if the API only ever returned one emoji.

* Modify and add unit tests for generateMessage

Unit tests have been modified so that the they work with the new string in the reminder message that includes the "not coming" emoji and respective text. Unit tests have also been added specifically for the new functionality, testing how the function handles cases such as duplicate emojis being passed

* Create independent function for emoji generation

Modified the generateMessage function to accept emojis as a parameter and removed the code that generates the emoji pair for the coming/not coming reactions in the message. Instead, a function has been created that returns a pair of non-duplicate emojis for use by generateMessage. The generateMessage call in the checkForEvents function has been modified to have the emojis parameter, but emojis will only be passed if the message being generated required them.

Since the emojis used in generated messages will soon need to be known by functions other than generateMessage (e.g. for reaction seeding), a function to generate emojis has been created and parameters in generateMessage modified to accept emojis through this function

* Modify unit tests for generateEmojiPair

The unit tests for generateMessage have been modified to work with the new generateEmojiFunction. Additionally, the emoji tests that were for generateMessage have now been modified to test the generateEmojiPair function

* Comment to clarify loop in generateEmojiPair

Added a comment to clarify the purpose of the loop in the generateEmojiPair function

* Update npm packages

* Modify getEmojiPair to return set emojis on fail

The getEmojiPair function has been modified so that instead of rejecting the promise after five attempts at getting a unique emoji pair, it will return a pair of hard-coded emojis: [":white_check_mark:", ":x:"] currently.

While the chance of the function failing when the Slack API is running is astronomically low, this implementation will ensure that the function will also function if the API fails to retrieve emojis.

* Reformat code

* Modify emoji duplicate checking loop

Modified the emoji duplicate checking loop in generateEmojiPair so that a duplicate flag is not needed.

Co-authored-by: Chris Sankey <csankey6@gmail.com>
Co-authored-by: Chris Sankey <5505109+chrissank@users.noreply.github.com>