Modify coming/not coming message string creation
The creation of the "React with x if you are coming/not" section of the message string in generateMessage has been modified so that each emoji is not first assigned a variable before being used to build the string.

Having variables are unnecessary for this section of the code as it at the very most provides a bit more readable code.