Modify if condition for seedMessageReactions
Modified the if condition that iterates through the message responses and seeds them with emojis. Previously, the condition was checking if the emojiPair array had a length property that's not zero, which would only exist if the array was initialized properly and had elements. When emojiPair is not needed, the generateMessage function initializes the variable to undefined, which has no length property, and thus would throw an error.

While this doesn't cause any issues with the operation of the function, it does throw TypeErrors, which are quite problematic.